+++ src/utils.js
@@ -1,6 +1,31 @@
 import { PLAN_COMPTABLE } from './plan-comptable.js';
 
 /**
+ * Affiche un message d'erreur dans l'élément HTML avec l'ID 'error-message'.
+ *
+ * @param {string} message - Le message d'erreur à afficher.
+ */
+function displayErrorMessage(message) {
+    const errorMessageElement = document.getElementById('error-message');
+    errorMessageElement.textContent = message;
+    errorMessageElement.style.display = 'block';
+}
+
+/**
+ * Gère une erreur en affichant un message d'erreur et en enregistrant l'erreur dans la console.
+ *
+ * @param {string} message - Le message d'erreur à afficher.
+ * @param {Object} line - La ligne de données associée à l'erreur.
+ * @throws {Error} - Lance une nouvelle erreur avec le message fourni.
+ */
+export function handleError(message, line) {
+    displayErrorMessage(message);
+    console.error(`Erreur: ${message} - Ligne: ${JSON.stringify(line)}`);
+    throw new Error(message);
+}
+
+
+/**
  * Trouve un compte comptable à partir de son numéro ou de son libellé.
  * @param {Object} param - Les paramètres de recherche.
  * @param {string} param.compte - Le numéro de compte.
